% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-cutpoints.R
\name{adaptCutpoint3groups}
\alias{adaptCutpoint3groups}
\title{A function used when two sets of cutpoints are defined
This function generates ROC Plots, Kaplan-Meier plots and histograms for two selected cutpoint values.}
\usage{
adaptCutpoint3groups(setCutpoint, df, time, event, biomarker)
}
\arguments{
\item{setCutpoint}{cutpoint variable}

\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
a folders for each biomarker with results
}
\description{
A function used when two sets of cutpoints are defined
This function generates ROC Plots, Kaplan-Meier plots and histograms for two selected cutpoint values.
}
