% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cutPMethod}
\alias{cutPMethod}
\title{Calculate cutpoints with cutp method.}
\usage{
cutPMethod(df, time, event, biomarker)
}
\arguments{
\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
Csv result table, histogram, Kaplan-Meier plot, heatmap
}
\description{
This function calculates the best cutpoint for a continuous variable with coxph and survfit model.
}
