% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-cutpoints.R
\name{defaultAnalyses}
\alias{defaultAnalyses}
\title{A function used when both sets of cutpoints are NULL
This function calculates the best cutpoint using cutp, maxstat, ROC and Youden methods, generates ROC Plots, Kaplan-Meier plots and histograms for 2 and 3 groups.}
\usage{
defaultAnalyses(df, time, event, biomarker)
}
\arguments{
\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
2 folders for each biomarker with results
}
\description{
A function used when both sets of cutpoints are NULL
This function calculates the best cutpoint using cutp, maxstat, ROC and Youden methods, generates ROC Plots, Kaplan-Meier plots and histograms for 2 and 3 groups.
}
