% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-cutpoints.R
\name{evaluateCutpoints}
\alias{evaluateCutpoints}
\title{A function creating folders for each biomarker in a data table with subfolders that include results of the analysis.}
\usage{
evaluateCutpoints(
  mainDir,
  file,
  resultsDirName,
  biomarkerList,
  time,
  event,
  setCutpoint,
  setCutpoint2,
  df
)
}
\arguments{
\item{time}{Time variable}

\item{event}{Event variable}

\item{df}{input data frame}

\item{biomarker}{Biomarker variable}

\item{cutpoints}{cutpoints variable}
}
\value{
results from a selected analysis
}
\description{
A function creating folders for each biomarker in a data table with subfolders that include results of the analysis.
}
