% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-cutpoints.R
\name{mainFunction}
\alias{mainFunction}
\title{A function that decides which type of analysis to use.}
\usage{
mainFunction(cutpoints, df, time, event, biomarker)
}
\arguments{
\item{cutpoints}{cutpoints variable}

\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
results from a selected analysis
}
\description{
A function that decides which type of analysis to use.
}
