% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{maxstatMethod}
\alias{maxstatMethod}
\title{Calculate cutpoints with maxstat method.}
\usage{
maxstatMethod(df, time, event, biomarker)
}
\arguments{
\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
Csv result table, histogram, Kaplan-Meier plot, Standarized log-rank statistics
}
\description{
This function calculates the best cutpoint for a continuous variable using maximally selected rank statistics.
}
