% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{oneCutpointAdapt}
\alias{oneCutpointAdapt}
\title{Calculate statistics for one selected cutpoint.}
\usage{
oneCutpointAdapt(setCutpoint, df, time, event, biomarker)
}
\arguments{
\item{setCutpoint}{cutpoint variable}

\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
Csv result table, histogram, Kaplan-Meier, Standarized log-rank statistics plots
}
\description{
This function calculates statistics if one set of cutpoints is selected.
}
