% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ROCMethod}
\alias{ROCMethod}
\title{Calculate cutpoints with ROC method.}
\usage{
ROCMethod(df, time, event, biomarker)
}
\arguments{
\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
ROC plot, youden Plot, a csv result table
}
\description{
This function calculates the best cutpoint using ROC method, generates a ROC Plot and histogram.
}
