% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{rolrMethod}
\alias{rolrMethod}
\title{Calculate cutpoints with maxstat method.}
\usage{
rolrMethod(setCutpoint, df, time, event, biomarker)
}
\arguments{
\item{setCutpoint}{cutpoint variable}

\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
Csv result table, histogram, Kaplan-Meier plots
}
\description{
This function calculates two cutpoints for a continuous variable using hierarchical method.
}
