% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{youdenMethod}
\alias{youdenMethod}
\title{Calculate cutpoints with Youden method.}
\usage{
youdenMethod(df, time, event, biomarker)
}
\arguments{
\item{df}{input data frame}

\item{time}{Time variable}

\item{event}{Event variable}

\item{biomarker}{Biomarker variable}
}
\value{
ROC plot, youden Plot, a csv result table
}
\description{
This function calculates the best cutpoint using Youden method, generates a ROC Plot and histogram.
}
